+++ anvolt/models/__init__.py
@@ -1,2 +1,3 @@
 from .response import Responses
-from .enums import Route, TwitchModels
+from .enums import Route, TwitchModels, MusicEnums, MusicPlatform, MusicPropertiesEnums
+from .player import MusicProperty, QueueSession