+++ anvolt/models/enums.py
@@ -1,4 +1,4 @@
-from enum import Enum
+from enum import Enum, IntEnum
 
 
 class Route(Enum):
@@ -64,3 +64,34 @@ class TwitchModels(Enum):
     STREAM_TAG_IDS = "tag_ids"
     STREAM_TAGS = "tags"
     STREAM_IS_MATURE = "is_mature"
+
+
+class MusicPropertiesEnums(Enum):
+    AUDIO_URL = 0
+    VIDEO_ID = 1
+    VIDEO_URL = 2
+    TITLE = 3
+    DURATION = 4
+    CURRENT_DURATION = 5
+    THUMBNAILS = 6
+    IS_LIVE = 7
+    REQUESTER = 8
+    VOLUME = 9
+    START_TIME = 10
+    LOOP = 11
+
+
+class MusicEnums(IntEnum):
+    """An enumeration class defining various music player settings, including loop options."""
+
+    NO_LOOPS = 0
+    LOOPS = 1
+    QUEUE_LOOPS = 2
+
+
+class MusicPlatform(IntEnum):
+    """An enumeration class defining the available platforms supported by anvolt.py"""
+
+    YOUTUBE_URL = 0
+    YOUTUBE_QUERY = 1
+    SOUNDCLOUD = 2